Change page sections implementation
Firstly, it is more DRY to have a single class (.page-section__heading
in our case) to set styles for all page headings at once than to set
them for every section separately.

Secondly, it is also necessary to change rules order in _general.scss
file to let the specificity correct styles for general BEM blocks.

In the currrent case the general .heading block has the value of margin
css property set to zero but every .page-section heading should be equal
to 48px.